Commented and explained the script
Added a more clear explanation of how to use the script, what systems it can be used on, and how it works. Also cleaned the script up a bit and removed "sudo" in front of the smartctl commands as it was unnecessary (since the script is supposed to be run by root or sudo user in the first place).